backend/requirements.txt
fastapi
uvicorn
pinecone
sentence-transformers
google-generativeai
langchain-community
langchain-pinecone
python-dotenv